bot/requirements.txt
aiohttp==3.8.5
aioprocessing==2.0.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
bitarray==2.8.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
cytoolz==0.12.2
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-retry==0.1.18
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fonttools==4.42.1
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lru-dict==1.2.0
matplotlib==3.7.2
multiaddr==0.0.9
multicall==0.7.4
multidict==6.0.4
netaddr==0.8.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parsimonious==0.8.1
Pillow==10.0.0
protobuf==3.19.5
pycryptodome==3.18.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==2.0.1
rpds-py==0.9.2
six==1.16.0
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
varint==1.0.2
web3==5.31.4
websockets==11.0.3
yarl==1.9.2
